+++ packages/components/src/form/FileUpload/FileUpload.stories.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { Box } from '@mantine/core';
-import { FileUpload, FILE_UPLOAD_DEFAULT_PROPS } from './FileUpload';
-import mdx from './FileUpload.mdx';
 import { MIME_TYPES } from '@mantine/dropzone';
 import { CloudUploadIcon } from '@bubbles-ui/icons/outline';
+import { FileUpload, FILE_UPLOAD_DEFAULT_PROPS } from './FileUpload';
+import mdx from './FileUpload.mdx';
 
 export default {
   title: 'Molecules/Form/FileUpload',
@@ -20,17 +20,40 @@ export default {
   argTypes: {
     onChange: { action: 'File dropped' },
     accept: { options: MIME_TYPES, control: { type: 'multi-select' } },
+    status: { options: ['error', 'loading', 'success'], control: { type: 'select' } },
   },
 };
 
-const Template = ({ children, ...props }) => {
-  return <FileUpload {...props}>{children}</FileUpload>;
-};
+const Template = ({ children, ...props }) => <FileUpload {...props}>{children}</FileUpload>;
 
 export const Playground = Template.bind({});
 
 Playground.args = {
   ...FILE_UPLOAD_DEFAULT_PROPS,
+  icon: <CloudUploadIcon height={32} width={32} />,
+  title: 'Click to browse your file',
+  subtitle: 'or drop here a file from your computer',
+  errorMessage: { title: 'Error', message: 'File was rejected' },
+};
+
+export const WithLoaders = Template.bind({});
+
+WithLoaders.args = {
+  ...FILE_UPLOAD_DEFAULT_PROPS,
+  initialFiles: [
+    {
+      name: 'My document.docx',
+      status: 'success',
+    },
+    {
+      name: 'My awesome image.svg',
+      status: 'loading',
+    },
+    {
+      name: 'My boring pdf.pdf',
+      status: 'error',
+    },
+  ],
   icon: <CloudUploadIcon height={32} width={32} />,
   title: 'Click to browse your file',
   subtitle: 'or drop here a file from your computer',