+++ packages/icons/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @bubbles-ui/icons
 
+## 1.2.125
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.124
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.123
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.122
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.121
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.120
+
+### Patch Changes
+
+- Versions bump
+
+## 1.2.119
+
+### Patch Changes
+
+- Versions bump
+
 ## 1.2.118
 
 ### Patch Changes